fixup: prompt failure report when the worker dies during model load + Florence dtype
A worker that exits during load (bad path, missing download) never set the ready
event, so _ensure_caption_worker blocked its background thread for the full 600 s
and then raised a stale error dialog long after the UI had already recovered via
the exit handler. The exit handler now SETS the event with warm still False - the
wait releases immediately and the warm check reports the failure once, promptly.

Also torch_dtype instead of dtype in the Florence load: the pinned transformers
reads torch_dtype, and the unknown kwarg fell through silently - Florence loaded
fp32 at double VRAM with nothing reporting it.

+++ src/fizgig/scripts/batch_caption.py
@@ -103,10 +103,12 @@ def _load_florence(model_name: str, revision: str | None, code_revision: str | N
         kwargs["code_revision"] = code_revision
     processor = from_pretrained_cache_first(AutoProcessor, model_name, **kwargs)
     dtype = torch.float16 if device == "cuda" else torch.float32
+    # torch_dtype, not dtype: the pinned transformers reads torch_dtype; an unknown kwarg
+    # falls through silently and Florence loads fp32 (double VRAM), which nothing reports.
     model = from_pretrained_cache_first(
         AutoModelForCausalLM,
         model_name,
-        dtype=dtype,
+        torch_dtype=dtype,
         attn_implementation="eager",
         **kwargs,
     ).to(device)
